# services/jurisdiction-rules-svc/openapi.yaml
openapi: "3.1.0"
info:
  title: Jurisdiction Rules Service
  version: "1.1.0"
  description: |
    Authoritative registry of jurisdiction identities and their effective-dated
    applicability rules (03-microservices.md §8.2).

    **Governance invariants:**
    - `rule_payload` holds APPLICABILITY METADATA ONLY — never computation values.
      Thresholds, rates and bands belong to the Tax and Payroll services (OQ-1, Model B).
    - Every type/status/domain field is a VARCHAR tag stored as data. New jurisdiction
      types, rule domains and status values are added by data migration, never by a
      code change (OQ-3).
    - Nothing is deleted. Jurisdictions are deactivated with `active_flag` + `effective_to`;
      rules are end-dated and moved to SUPERSEDED or RETIRED.
    - Historical actions must always be explainable against the rule set active at the
      time of execution, so point-in-time reads keep working after a jurisdiction is
      deactivated and after a rule is superseded.
    - `legal_drift_state` transitions are append-only. The column carries the current
      value; `/v1/rules/{id}/drift-events` carries the history (OQ-4).
    - Jurisdiction and rule data are platform-wide reference data. There is no
      `tenant_id` and no row-level security on these tables.

    **Read vs runtime resolution.** `/rules` is the raw per-jurisdiction, effective-dated
    view — the audit and replay surface, one jurisdiction only. `/rule-pack` is the
...

...

...
servers:
...
paths:
...
  /v1/admin/rollouts:
    post:
      tags: [Rollout]
      summary: Add a jurisdiction pack family to the rollout portfolio (PLANNED)
      description: |
        Action JURISDICTION_ROLLOUT_CREATE. Body: family_ref (for example gb or us.state.ca), display_name, layer (GLOBAL_REFERENCE, GLOBAL_CORE,
        REGIONAL_FRAMEWORK, COUNTRY, SUBDIVISION), jurisdiction_code (required for COUNTRY and SUBDIVISION), owner, support_owner?, scope_note?.
        Naming a family here does NOT mean certified rules exist for it (s32).
      operationId: createRollout
      security: [{ principalHeader: [] }]
      parameters: [{ $ref: '#/components/parameters/PrincipalID' }]
      responses: { "201": { description: Created. }, "400": { $ref: '#/components/responses/BadRequest' }, "409": { description: One rollout per family_ref. } }
    get:
      tags: [Rollout]
      summary: List rollouts (optional status filter)
      description: Action JURISDICTION_ROLLOUT_VIEW.
      operationId: listRollouts
      security: [{ principalHeader: [] }]
      parameters: [{ $ref: '#/components/parameters/PrincipalID' }, { name: status, in: query, required: false, schema: { type: string, enum: [PLANNED, AUTHORING, READY, LAUNCHED, SUSPENDED, RETIRED] } }]
      responses: { "200": { description: Rollouts. } }
  /v1/admin/rollouts:seed-portfolio:
    post:
      tags: [Rollout]
      summary: Seed the s32 initial portfolio as PLANNED entries with no owner (idempotent)
      description: Action JURISDICTION_ROLLOUT_SEED. States intent only; a PLANNED entry is not support.
      operationId: seedRolloutPortfolio
      security: [{ principalHeader: [] }]
      parameters: [{ $ref: '#/components/parameters/PrincipalID' }]
      responses: { "200": { description: Counts of created and already present entries. } }
  /v1/admin/rollouts/{rollout_id}:
    get:
      tags: [Rollout]
      summary: A rollout with its checklist state, expert decisions, packs, history and what blocks each next step
      description: Action JURISDICTION_ROLLOUT_VIEW.
      operationId: getRollout
      security: [{ principalHeader: [] }]
      parameters: [{ $ref: '#/components/parameters/PrincipalID' }, { name: rollout_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: Detail. }, "404": { description: rollout_not_found. } }
  /v1/admin/rollouts/{rollout_id}/owner:
    put:
      tags: [Rollout]
      summary: Assign the accountable owner and support owner (DOR_09)
      description: Action JURISDICTION_ROLLOUT_SET_OWNER. A principal who already attested to or approved the rollout cannot become its owner. Not allowed once LAUNCHED or RETIRED.
      operationId: setRolloutOwner
      security: [{ principalHeader: [] }]
      parameters: [{ $ref: '#/components/parameters/PrincipalID' }, { name: rollout_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: Updated. }, "400": { $ref: '#/components/responses/BadRequest' }, "403": { description: segregation_of_duties. }, "404": { description: rollout_not_found. } }
  /v1/admin/rollouts/{rollout_id}/packs:
    put:
      tags: [Rollout]
      summary: Link the packs that make up the rollout (additive, idempotent)
      description: Action JURISDICTION_ROLLOUT_LINK_PACKS. Body pack_refs (1 to 50).
      operationId: linkRolloutPacks
      security: [{ principalHeader: [] }]
      parameters: [{ $ref: '#/components/parameters/PrincipalID' }, { name: rollout_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: Linked pack refs. }, "404": { description: rollout_not_found or pack_not_found. } }
  /v1/admin/rollouts/{rollout_id}/attestations:
    post:
      tags: [Rollout]
      summary: Record a Definition of Ready (DOR_01..10) or Definition of Done (DOD_01..11) answer
      description: |
        Action JURISDICTION_ROLLOUT_ATTEST. Body: checklist (READY or DONE), item_code, met, evidence_ref (required when met). The attester cannot be the
        rollout owner. Append-only: the latest answer for an item is current, so an item can be withdrawn by recording met false.
      operationId: attestRollout
      security: [{ principalHeader: [] }]
      parameters: [{ $ref: '#/components/parameters/PrincipalID' }, { name: rollout_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "201": { description: Recorded. }, "400": { $ref: '#/components/responses/BadRequest' }, "403": { description: segregation_of_duties. }, "404": { description: rollout_not_found. } }
  /v1/admin/rollouts/{rollout_id}/expert-approvals:
    post:
      tags: [Rollout]
      summary: Record a qualified local expert, legal or tax decision
      description: 'Action JURISDICTION_ROLLOUT_EXPERT_APPROVE. Body: qualification, scope, decision (APPROVE or REJECT), notes?. The expert cannot be the rollout owner. An expert''s latest decision counts.'
      operationId: expertApproveRollout
      security: [{ principalHeader: [] }]
      parameters: [{ $ref: '#/components/parameters/PrincipalID' }, { name: rollout_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "201": { description: Recorded. }, "400": { $ref: '#/components/responses/BadRequest' }, "403": { description: segregation_of_duties. }, "404": { description: rollout_not_found. } }
  /v1/admin/rollouts/{rollout_id}/transition:
    post:
      tags: [Rollout]
      summary: Move a rollout (AUTHORING, READY, LAUNCHED, SUSPENDED, RETIRED)
      description: |
        Action JURISDICTION_ROLLOUT_TRANSITION. Body: to, reason (required for SUSPENDED, RETIRED and for READY back to AUTHORING).
        READY needs the whole Definition of Ready. LAUNCHED additionally needs the whole Definition of Done, at least one approving qualified expert, a linked pack with a
        RELEASED version, and a launcher who is neither the owner nor an expert on the rollout. A refused move returns 409 with EVERY blocker. The database enforces the same gates.
      operationId: transitionRollout
      security: [{ principalHeader: [] }]
      parameters: [{ $ref: '#/components/parameters/PrincipalID' }, { name: rollout_id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: Moved. }, "400": { $ref: '#/components/responses/BadRequest' }, "404": { description: rollout_not_found. }, "409": { description: 'rollout_blocked, with blockers.' } }
  /v1/rollout-checklists:
    get:
      tags: [Rollout]
      summary: The Definition of Ready (s35) and Definition of Done (s36) items
      description: Action JURISDICTION_ROLLOUT_VIEW.
      operationId: rolloutChecklists
      security: [{ principalHeader: [] }]
      parameters: [{ $ref: '#/components/parameters/PrincipalID' }]
      responses: { "200": { description: The two checklists. } }
  /v1/jurisdiction-support/{jurisdiction_code}:
    get:
      tags: [Rollout]
      summary: Is this jurisdiction supported for production? (never guesses)
      description: |
        Action JURISDICTION_SUPPORT_VIEW. SUPPORTED only when a LAUNCHED rollout names the jurisdiction AND a linked pack has a RELEASED version covering it. Otherwise
        NOT_YET_SUPPORTED, SUSPENDED, NO_RELEASED_PACK (launched but nothing released covers it, for example after a withdrawal) or NO_ROLLOUT, each with an explanation (JUR-NEG-22).
        Matches the rollout's jurisdiction_code exactly: a subdivision is not supported because its country is.
      operationId: jurisdictionSupport
      security: [{ principalHeader: [] }]
      parameters: [{ $ref: '#/components/parameters/PrincipalID' }, { name: jurisdiction_code, in: path, required: true, schema: { type: string, maxLength: 32 } }]
      responses: { "200": { description: Support status. } }

  # ── Health ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    principalHeader:
      type: apiKey
      in: header
      name: X-Principal-Id
      description: |
        The acting principal, set by the gateway's ForwardAuth middleware after it
        verifies the caller's identity envelope against identity-svc's JWKS. Callers
        do not set this themselves — a request that reaches the service without it is
        rejected with `401 missing_principal` rather than attributed to the platform.
  parameters:
...
    PrincipalID:
      name: X-Principal-Id
      in: header
      required: true
      description: See the `principalHeader` security scheme.
      schema: { type: string }
...
  responses:
    BadRequest:
      description: |
        Malformed or incomplete request. Codes include `missing_field`,
        `invalid_request_body`, `invalid_effective_at`, `invalid_effective_period`,
        `invalid_limit`, `invalid_offset`, `invalid_rule_status`,
        `invalid_rule_payload`, `invalid_status`, `invalid_drift_state`.
        An unknown JSON field is rejected rather than ignored.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      required: [type, title, status, detail, instance]
      description: |
        RFC 9457 Problem Details format. Clients can switch on `type` or `title`.
        The `extension` fields provide machine-readable codes.
      properties:
        type:
          type: string
          format: uri
          description: URI identifying the problem type.
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          description: HTTP status code generated by the origin server.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: URI reference identifying the specific occurrence.
        code:
          type: string
          description: Stable machine-readable code for client logic (e.g., "missing_field").
        field:
          type: string
          description: The offending field, when code is "missing_field".
        jurisdiction_id:
          type: string
          format: uuid
          description: The jurisdiction ID, when relevant.
...